requirements.txt
langchain
langchain_gigachat
langgraph
langchain_community
duckduckgo-search
python-dotenv
pillow
